refactor(kitchenpos): dao 인터페이스 분리
* DI하여 테스트하기 위해 인터페이스로 분리
