+++ src/main/java/camp/nextstep/edu/kitchenpos/bo/MenuGroupBo.java
@@ -9,6 +9,7 @@ import java.util.List;
 
 @Component
 public class MenuGroupBo {
+
     private final MenuGroupDao menuGroupDao;
 
     public MenuGroupBo(final MenuGroupDao menuGroupDao) {
